+++ src/Callisto/Controls/WatermarkTextBox/WatermarkTextBox.cs
@@ -15,6 +15,7 @@
 //
 
 using Callisto.Controls.Common;
+using System;
 using Windows.ApplicationModel.Resources;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
@@ -22,6 +23,7 @@ using Windows.UI.Xaml.Input;
 
 namespace Callisto.Controls
 {
+    [Obsolete("Windows 8.1 now provides this functionality in the XAML framework itself as PlaceholderText on TextBox.")]
     [TemplatePart(Name = WatermarkTextBox.PART_ELEMENT_CONTENT_NAME, Type = typeof(ContentControl))]
     public sealed class WatermarkTextBox : TextBox
     {
@@ -163,6 +165,9 @@ namespace Callisto.Controls
             base.OnApplyTemplate();
 
             ElementContent = (TextBlock)GetTemplateChild(PART_ELEMENT_CONTENT_NAME);
+
+            // if the PlaceholderText value (new in Win8.1) is set, we will use that as Watermark
+            if (!string.IsNullOrEmpty(PlaceholderText)) Watermark = PlaceholderText;
 
             OnWatermarkChanged();