+++ changelog.md
@@ -1,7 +1,12 @@
 # Changelog
 
-## 1.3.1 (25-JUN-2013)
-* Fix #214: Removed explicit font settings on WatermarkTextBox to allow inheritance to flow for Font properties
+## 1.4.0 (25-OCT-2013)
+* BREAKING CHANGES!!!
+* Starting in 1.4.0, Callisto will only support Windows 8.1 applications
+* Deprecated: Flyout, Menu, SettingsFlyout, WatermarkTextBox
+* MODIFIED: AppSettings/SettingsManagement/AppManifestHelper APIs; changed to ThemeResource in generic.xaml
+* NEW: Blend design-time support
+* NuGet only supports netcore451 now going forward
 
 ## 1.3.0 (21-MAY-2013)
 * Fix #204: Fixed memory leak in CustomDialog (thanks kiwidev!)